[feat] 알림 사운드 설정 및 위치(main -> splash) 변경
